4545 Move platform-ai-skill to top-level EE platform
Lift platform-ai-skill out of the platform-ai-agent nesting and relocate
to server/ee/libs/platform/platform-ai/platform-ai-skill/, matching the
EE convention used by platform-ai-llm-usage / platform-ai-tool-usage.

The pre-existing platform-ai-skill-remote-client (already under EE) is
moved alongside its siblings. External call sites updated:
- server-app, worker-app: dependency paths
- modules/components/ai/agent/utils: import rewrites

Package rename: com.bytechef.platform.ai.skill →
com.bytechef.ee.platform.ai.skill, matching the existing
com.bytechef.ee.platform.ai.skill.remote.client convention.
