+++ server/ee/apps/worker-app/build.gradle.kts
@@ -62,7 +62,7 @@ dependencies {
     }
 
     implementation(project(":server:ee:libs:atlas:atlas-execution:atlas-execution-remote-client"))
-    implementation(project(":server:ee:libs:platform:platform-ai:platform-ai-agent:platform-ai-skill:platform-ai-skill-remote-client"))
+    implementation(project(":server:ee:libs:platform:platform-ai:platform-ai-skill:platform-ai-skill-remote-client"))
     implementation(project(":server:ee:libs:config:observability-config"))
     implementation(project(":server:ee:libs:core:discovery:discovery-metadata-api"))
     implementation(project(":server:ee:libs:core:discovery:discovery-redis"))
