+++ server/ee/libs/platform/platform-ai/platform-ai-skill/platform-ai-skill-api/build.gradle.kts
