+++ server/ee/libs/platform/platform-ai/platform-ai-skill/platform-ai-skill-api/src/main/java/com/bytechef/ee/platform/ai/skill/domain/AiSkill.java
@@ -1,0 +1,172 @@
+/*
+ * Copyright 2025 ByteChef
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bytechef.ee.platform.ai.skill.domain;
+
+import com.bytechef.file.storage.domain.FileEntry;
+import java.time.Instant;
+import java.util.Objects;
+import org.jspecify.annotations.Nullable;
+import org.springframework.data.annotation.CreatedBy;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.annotation.LastModifiedBy;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.annotation.Version;
+import org.springframework.data.relational.core.mapping.Column;
+import org.springframework.data.relational.core.mapping.Table;
+
+/**
+ * @author Ivica Cardic
+ */
+@Table("ai_skill")
+public final class AiSkill {
+
+    @Id
+    private Long id;
+
+    private String name;
+
+    @Nullable
+    private String description;
+
+    @Column("skill_file")
+    private FileEntry skillFile;
+
+    @CreatedDate
+    private Instant createdDate;
+
+    @CreatedBy
+    private String createdBy;
+
+    @LastModifiedDate
+    private Instant lastModifiedDate;
+
+    @LastModifiedBy
+    private String lastModifiedBy;
+
+    @Version
+    private int version;
+
+    public AiSkill() {
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Nullable
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(@Nullable String description) {
+        this.description = description;
+    }
+
+    public FileEntry getSkillFile() {
+        return skillFile;
+    }
+
+    public void setSkillFile(FileEntry skillFile) {
+        this.skillFile = skillFile;
+    }
+
+    public Instant getCreatedDate() {
+        return createdDate;
+    }
+
+    public void setCreatedDate(Instant createdDate) {
+        this.createdDate = createdDate;
+    }
+
+    public String getCreatedBy() {
+        return createdBy;
+    }
+
+    public void setCreatedBy(String createdBy) {
+        this.createdBy = createdBy;
+    }
+
+    public Instant getLastModifiedDate() {
+        return lastModifiedDate;
+    }
+
+    public void setLastModifiedDate(Instant lastModifiedDate) {
+        this.lastModifiedDate = lastModifiedDate;
+    }
+
+    public String getLastModifiedBy() {
+        return lastModifiedBy;
+    }
+
+    public void setLastModifiedBy(String lastModifiedBy) {
+        this.lastModifiedBy = lastModifiedBy;
+    }
+
+    public int getVersion() {
+        return version;
+    }
+
+    public void setVersion(int version) {
+        this.version = version;
+    }
+
+    @Override
+    public boolean equals(Object object) {
+        if (this == object) {
+            return true;
+        }
+
+        if (!(object instanceof AiSkill aiSkill)) {
+            return false;
+        }
+
+        return Objects.equals(id, aiSkill.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return getClass().hashCode();
+    }
+
+    @Override
+    public String toString() {
+        return "AiSkill{" +
+            "id=" + id +
+            ", name='" + name + '\'' +
+            ", description='" + description + '\'' +
+            ", skillFile='" + skillFile + '\'' +
+            ", createdDate=" + createdDate +
+            ", createdBy='" + createdBy + '\'' +
+            ", lastModifiedDate=" + lastModifiedDate +
+            ", lastModifiedBy='" + lastModifiedBy + '\'' +
+            ", version=" + version +
+            '}';
+    }
+}
