+++ server/ee/libs/platform/platform-ai/platform-ai-skill/platform-ai-skill-api/src/main/java/com/bytechef/ee/platform/ai/skill/service/AiSkillService.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2025 ByteChef
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bytechef.ee.platform.ai.skill.service;
+
+import com.bytechef.ee.platform.ai.skill.domain.AiSkill;
+import java.util.List;
+import org.jspecify.annotations.Nullable;
+
+/**
+ * @author Ivica Cardic
+ */
+public interface AiSkillService {
+
+    AiSkill createAiSkill(AiSkill aiSkill);
+
+    void deleteAiSkill(long id);
+
+    boolean existsByName(String name);
+
+    AiSkill getAiSkill(long id);
+
+    List<AiSkill> getAiSkills();
+
+    AiSkill updateAiSkill(long id, String name, @Nullable String description);
+}
