+++ server/ee/libs/platform/platform-ai/platform-ai-skill/platform-ai-skill-file-storage/platform-ai-skill-file-storage-api/src/main/java/com/bytechef/ee/platform/ai/skill/file/storage/AiSkillFileStorage.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2025 ByteChef
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bytechef.ee.platform.ai.skill.file.storage;
+
+import com.bytechef.file.storage.domain.FileEntry;
+
+/**
+ * @author Ivica Cardic
+ */
+public interface AiSkillFileStorage {
+
+    void deleteAiSkillFile(FileEntry skillFile);
+
+    byte[] readAiSkillFileBytes(FileEntry skillFile);
+
+    FileEntry storeAiSkillFile(String filename, byte[] bytes);
+}
