+++ server/ee/libs/platform/platform-ai/platform-ai-skill/platform-ai-skill-graphql/build.gradle.kts
@@ -1,0 +1,6 @@
+dependencies {
+    implementation(project(":server:ee:libs:platform:platform-ai:platform-ai-skill:platform-ai-skill-api"))
+
+    implementation("org.springframework.boot:spring-boot-autoconfigure")
+    implementation("org.springframework.graphql:spring-graphql")
+}
