+++ server/ee/libs/platform/platform-ai/platform-ai-skill/platform-ai-skill-remote-client/build.gradle.kts
@@ -1,0 +1,6 @@
+dependencies {
+    implementation("org.springframework:spring-context")
+    implementation(project(":server:ee:libs:platform:platform-ai:platform-ai-skill:platform-ai-skill-api"))
+
+    implementation(project(":server:ee:libs:core:remote:remote-client"))
+}
