+++ server/ee/libs/platform/platform-ai/platform-ai-skill/platform-ai-skill-rest/build.gradle.kts
@@ -1,0 +1,6 @@
+dependencies {
+    implementation(project(":server:ee:libs:platform:platform-ai:platform-ai-skill:platform-ai-skill-api"))
+
+    implementation("org.springframework:spring-web")
+    implementation("org.springframework.boot:spring-boot-autoconfigure")
+}
