+++ server/ee/libs/platform/platform-ai/platform-ai-skill/platform-ai-skill-rest/src/main/java/com/bytechef/ee/platform/ai/skill/web/rest/AiSkillDownloadController.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2025 ByteChef
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bytechef.ee.platform.ai.skill.web.rest;
+
+import com.bytechef.ee.platform.ai.skill.facade.AiSkillFacade;
+import com.bytechef.ee.platform.ai.skill.facade.AiSkillFacade.AiSkillDownload;
+import org.springframework.core.io.ByteArrayResource;
+import org.springframework.core.io.Resource;
+import org.springframework.http.ContentDisposition;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * @author Ivica Cardic
+ */
+@RestController
+@RequestMapping("/api/ai/agent-skills")
+class AiSkillDownloadController {
+
+    private final AiSkillFacade aiSkillFacade;
+
+    AiSkillDownloadController(AiSkillFacade aiSkillFacade) {
+        this.aiSkillFacade = aiSkillFacade;
+    }
+
+    @GetMapping("/{id}/download")
+    ResponseEntity<Resource> downloadAiSkill(@PathVariable long id) {
+        AiSkillDownload download = aiSkillFacade.getAiSkillWithDownload(id);
+
+        String safeName = download.aiSkill()
+            .getName()
+            .replaceAll("[^a-zA-Z0-9._\\- ]", "_");
+
+        ContentDisposition contentDisposition = ContentDisposition.attachment()
+            .filename(safeName + ".skill")
+            .build();
+
+        return ResponseEntity.ok()
+            .contentType(MediaType.APPLICATION_OCTET_STREAM)
+            .header(HttpHeaders.CONTENT_DISPOSITION, contentDisposition.toString())
+            .body(new ByteArrayResource(download.bytes()));
+    }
+}
