+++ server/ee/libs/platform/platform-ai/platform-ai-skill/platform-ai-skill-service/src/main/java/com/bytechef/ee/platform/ai/skill/repository/AiSkillRepository.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2025 ByteChef
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bytechef.ee.platform.ai.skill.repository;
+
+import com.bytechef.ee.platform.ai.skill.domain.AiSkill;
+import org.springframework.data.repository.ListCrudRepository;
+import org.springframework.stereotype.Repository;
+
+/**
+ * @author Ivica Cardic
+ */
+@Repository
+public interface AiSkillRepository extends ListCrudRepository<AiSkill, Long> {
+
+    boolean existsByName(String name);
+}
