+++ server/ee/libs/platform/platform-ai/platform-ai-skill/platform-ai-skill-service/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+com.bytechef.ee.platform.ai.skill.config.AiSkillJdbcRepositoryConfiguration
