+++ server/ee/libs/platform/platform-ai/platform-ai-skill/platform-ai-skill-service/src/main/resources/config/liquibase/changelog/ai/skill/20260503000001_ai_skill_rename_table.xml
