+++ server/libs/modules/components/ai/agent/utils/build.gradle.kts
@@ -1,6 +1,6 @@
 dependencies {
     implementation(libs.org.springaicommunity.spring.ai.agent.utils)
-    implementation(project(":server:libs:platform:platform-ai:platform-ai-agent:platform-ai-skill:platform-ai-skill-api"))
+    implementation(project(":server:ee:libs:platform:platform-ai:platform-ai-skill:platform-ai-skill-api"))
     implementation(project(":server:libs:core:commons:commons-util"))
     implementation(project(":server:libs:core:file-storage:file-storage-api"))
     implementation(project(":server:libs:platform:platform-component:platform-component-api"))
