--- server/libs/platform/platform-ai/platform-ai-agent/platform-ai-skill/platform-ai-skill-file-storage/platform-ai-skill-file-storage-impl/build.gradle.kts
@@ -1,5 +1,0 @@
-dependencies {
-    implementation("org.springframework.boot:spring-boot-autoconfigure")
-    implementation(project(":server:libs:config:app-config"))
-    implementation(project(":server:libs:platform:platform-ai:platform-ai-agent:platform-ai-skill:platform-ai-skill-file-storage:platform-ai-skill-file-storage-api"))
-}
