--- server/libs/platform/platform-ai/platform-ai-agent/platform-ai-skill/platform-ai-skill-graphql/build.gradle.kts
@@ -1,6 +1,0 @@
-dependencies {
-    implementation(project(":server:libs:platform:platform-ai:platform-ai-agent:platform-ai-skill:platform-ai-skill-api"))
-
-    implementation("org.springframework.boot:spring-boot-autoconfigure")
-    implementation("org.springframework.graphql:spring-graphql")
-}
