--- server/libs/platform/platform-ai/platform-ai-agent/platform-ai-skill/platform-ai-skill-rest/build.gradle.kts
@@ -1,6 +1,0 @@
-dependencies {
-    implementation(project(":server:libs:platform:platform-ai:platform-ai-agent:platform-ai-skill:platform-ai-skill-api"))
-
-    implementation("org.springframework:spring-web")
-    implementation("org.springframework.boot:spring-boot-autoconfigure")
-}
