--- server/libs/platform/platform-ai/platform-ai-agent/platform-ai-skill/platform-ai-skill-service/src/main/java/com/bytechef/platform/ai/skill/config/AiSkillJdbcRepositoryConfiguration.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2025 ByteChef
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.bytechef.platform.ai.skill.config;
-
-import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
-import org.springframework.data.jdbc.repository.config.AbstractJdbcConfiguration;
-import org.springframework.data.jdbc.repository.config.EnableJdbcRepositories;
-
-/**
- * @author Ivica Cardic
- */
-@AutoConfiguration(afterName = "org.springframework.boot.data.jdbc.autoconfigure.DataJdbcRepositoriesAutoConfiguration")
-@ConditionalOnBean(AbstractJdbcConfiguration.class)
-@EnableJdbcRepositories(basePackages = "com.bytechef.platform.ai.skill.repository")
-class AiSkillJdbcRepositoryConfiguration {
-}
