--- server/libs/platform/platform-ai/platform-ai-agent/platform-ai-skill/platform-ai-skill-service/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-com.bytechef.platform.ai.skill.config.AiSkillJdbcRepositoryConfiguration
