+++ settings.gradle.kts
@@ -32,17 +32,6 @@ include("cli:commands:component:init:openapi")
 
 include("server:apps:server-app")
 
-include("server:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-skill:platform-ai-agent-skill-api")
-include("server:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-skill:platform-ai-agent-skill-file-storage:platform-ai-agent-skill-file-storage-api")
-include("server:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-skill:platform-ai-agent-skill-file-storage:platform-ai-agent-skill-file-storage-impl")
-include("server:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-skill:platform-ai-agent-skill-graphql")
-include("server:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-skill:platform-ai-agent-skill-rest")
-include("server:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-skill:platform-ai-agent-skill-service")
-include("server:ee:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-eval:platform-ai-agent-eval-api")
-include("server:ee:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-eval:platform-ai-agent-eval-file-storage:platform-ai-agent-eval-file-storage-api")
-include("server:ee:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-eval:platform-ai-agent-eval-file-storage:platform-ai-agent-eval-file-storage-impl")
-include("server:ee:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-eval:platform-ai-agent-eval-graphql")
-include("server:ee:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-eval:platform-ai-agent-eval-service")
 include("server:libs:platform:platform-ai:platform-ai-tool-api")
 include("server:libs:ai:mcp:mcp-server")
 include("server:libs:ai:mcp:mcp-server-configuration:mcp-server-configuration-graphql")
@@ -237,7 +226,6 @@ include("server:libs:modules:components:agile-crm")
 include("server:libs:modules:components:aha")
 include("server:libs:modules:components:ahrefs")
 include("server:libs:modules:components:ai:agent")
-include("server:libs:modules:components:ai:agent:chat-memory")
 include("server:libs:modules:components:ai:agent:chat-memory:chat-memory-builtin")
 include("server:libs:modules:components:ai:agent:chat-memory:chat-memory-cassandra")
 include("server:libs:modules:components:ai:agent:chat-memory:chat-memory-cosmosdb")
@@ -535,7 +523,7 @@ include("server:ee:libs:atlas:atlas-execution:atlas-execution-remote-client")
 include("server:ee:libs:ai:ai-copilot:ai-copilot-api")
 include("server:ee:libs:ai:ai-copilot:ai-copilot-rest")
 include("server:ee:libs:ai:ai-copilot:ai-copilot-service")
-include("server:ee:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-skill:platform-ai-agent-skill-remote-client")
+
 include("server:ee:libs:automation:automation-api-platform:automation-api-platform-configuration:automation-api-platform-configuration-api")
 include("server:ee:libs:automation:automation-api-platform:automation-api-platform-configuration:automation-api-platform-configuration-rest")
 include("server:ee:libs:automation:automation-api-platform:automation-api-platform-configuration:automation-api-platform-configuration-service")
@@ -609,6 +597,18 @@ include("server:ee:libs:embedded:embedded-workflow:embedded-workflow-execution:e
 include("server:ee:libs:embedded:embedded-workflow:embedded-workflow-execution:embedded-workflow-execution-rest")
 include("server:ee:libs:embedded:embedded-workflow:embedded-workflow-execution:embedded-workflow-execution-service")
 
+include("server:ee:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-eval:platform-ai-agent-eval-api")
+include("server:ee:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-eval:platform-ai-agent-eval-file-storage:platform-ai-agent-eval-file-storage-api")
+include("server:ee:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-eval:platform-ai-agent-eval-file-storage:platform-ai-agent-eval-file-storage-impl")
+include("server:ee:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-eval:platform-ai-agent-eval-graphql")
+include("server:ee:libs:platform:platform-ai:platform-ai-agent:platform-ai-agent-eval:platform-ai-agent-eval-service")
+include("server:ee:libs:platform:platform-ai:platform-ai-skill:platform-ai-skill-api")
+include("server:ee:libs:platform:platform-ai:platform-ai-skill:platform-ai-skill-file-storage:platform-ai-skill-file-storage-api")
+include("server:ee:libs:platform:platform-ai:platform-ai-skill:platform-ai-skill-file-storage:platform-ai-skill-file-storage-impl")
+include("server:ee:libs:platform:platform-ai:platform-ai-skill:platform-ai-skill-graphql")
+include("server:ee:libs:platform:platform-ai:platform-ai-skill:platform-ai-skill-remote-client")
+include("server:ee:libs:platform:platform-ai:platform-ai-skill:platform-ai-skill-rest")
+include("server:ee:libs:platform:platform-ai:platform-ai-skill:platform-ai-skill-service")
 include("server:ee:libs:platform:platform-api-connector:platform-api-connector-configuration:platform-api-connector-configuration-api")
 include("server:ee:libs:platform:platform-api-connector:platform-api-connector-configuration:platform-api-connector-configuration-graphql")
 include("server:ee:libs:platform:platform-api-connector:platform-api-connector-configuration:platform-api-connector-configuration-service")
